Backend/requirements.txt
fastapi
pydantic
openai
google-genai
python-dotenv
uvicorn
